refactor: move types in such a way they can be tree shaken
