+++ src/utils/default_snippet_variables.ts
@@ -1,10 +1,10 @@
 // @ts-ignore
 import default_snippet_variables_str from "inline:src/default_snippet_variables.json5";
-import json5 from "json5";
 import type { SnippetVariables } from "src/extension";
+// @ts-ignore
+import default_snippet_variables_json from "../default_snippet_variables.json5";
 
 export const DEFAULT_SNIPPET_VARIABLES_str: string =
     default_snippet_variables_str;
-export const DEFAULT_SNIPPET_VARIABLES: SnippetVariables = json5.parse(
-    default_snippet_variables_str,
-);
+export const DEFAULT_SNIPPET_VARIABLES: SnippetVariables =
+    default_snippet_variables_json;
